merge: resolve conflicts with origin main
Agent-Logs-Url: https://github.com/Masterleeaus/Worksuite-Saas---Project-Management-System_Laravel/sessions/bab7bd37-6756-49ab-ad5a-f93015994c1f

+++ Modules/CustomerModule/Database/Migrations/2026_04_10_700002_create_client_addresses_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
 
         Schema::create('client_addresses', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedBigInteger('client_id')->index();
-            $table->unsignedBigInteger('company_id')->nullable()->index();
+            $table->unsignedInteger('client_id')->index();
+            $table->unsignedInteger('company_id')->nullable()->index();
             $table->string('label', 100)->nullable();         // e.g. "Home", "Investment Property"
             $table->string('address_line_1', 255)->nullable();
             $table->string('address_line_2', 255)->nullable();
